fixup: hoist peer_check types into parent; downgrade error logs to debug
- Restructure unix/peer_check as a directory module: parent peer_check.rs
  owns the shared Allowlist + PeerRejected types, with platform-specific
  impls in peer_check/linux.rs and peer_check/macos.rs. Drops the duplicate
  enum definitions per thomaseizinger's review.
- Downgrade the read_allowlist_file failure logs from error to debug;
  these would otherwise hit Sentry on every misconfigured install. The
  one-shot startup info log for "file missing" stays at info.
- Document why we still validate the allowlist file's ownership/mode:
  without it, an attacker who can write to /etc/firezone/allowed-clients.conf
  alone could allowlist a root-owned interpreter (bash, python) and use
  it as a confused deputy. The check defends against that narrow path.
